restructuring package
